refactor: move form components to separate folders
