+++ src/components/ui/inputs/LocationPicker.tsx
@@ -2,7 +2,7 @@ import { PopUp } from "../../crm/AddCardCrm/PopUp/index";
 import { ArrowDownIcon } from "../icon/ArrowDownIcon";
 import { ArrowUpIcon } from "../icon/ArrowUpIcon";
 import { CurrentLocation } from "../../crm/AddCardCrm/PopUp/CurrentLocation";
-import { Location } from "../../crm/AddCardCrm/AddCardForm";
+import { Location } from "../../crm/AddCardForm";
 import { TextInput } from "./TextInput";
 
 interface PropsLocationPicker {
